requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
flask=2.1.2=pypi_0
flask-cors=3.0.10=pypi_0
keras=2.9.0=pypi_0
numpy=1.22.4=pypi_0
pillow=9.1.1=pypi_0
tensorflow=2.9.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0

